+++ backend/bots/examples/group_prize_bot/api/src/updates/join_group.rs
@@ -14,6 +14,7 @@ pub enum Response {
     AlreadyInGroup,
     GroupNotFound,
     GroupNotPublic,
+    FailedToPassGateCheck(String),
     ParticipantLimitReached,
     Blocked,
     UserSuspended,