+++ backend/canisters/group/impl/src/lib.rs
@@ -14,8 +14,8 @@ use std::collections::HashMap;
 use std::ops::Deref;
 use types::{
     Avatar, CanisterId, ChatId, Cryptocurrency, Cycles, EventIndex, FrozenGroupInfo, GroupCanisterGroupChatSummary,
-    GroupPermissions, GroupRules, GroupSubtype, MessageIndex, Milliseconds, Notification, TimestampMillis, Timestamped, UserId,
-    Version, MAX_THREADS_IN_SUMMARY,
+    GroupPermissions, GroupRules, GroupSubtype, JoinGroupGate, MessageIndex, Milliseconds, Notification, TimestampMillis,
+    Timestamped, UserId, Version, MAX_THREADS_IN_SUMMARY,
 };
 use utils::env::Environment;
 use utils::regular_jobs::RegularJobs;
@@ -253,6 +253,8 @@ struct Data {
     pub frozen: Timestamped<Option<FrozenGroupInfo>>,
     pub timer_jobs: TimerJobs<TimerJob>,
     pub date_last_pinned: Option<TimestampMillis>,
+    #[serde(default)]
+    pub gate: Option<JoinGroupGate>,
 }
 
 fn initialize_ledger_ids() -> HashMap<Cryptocurrency, CanisterId> {
@@ -302,6 +304,7 @@ impl Data {
         proposals_bot_user_id: UserId,
         test_mode: bool,
         permissions: Option<GroupPermissions>,
+        gate: Option<JoinGroupGate>,
     ) -> Data {
         let participants = Participants::new(creator_principal, creator_user_id, now);
         let events = ChatEvents::new_group_chat(chat_id, name.clone(), description.clone(), creator_user_id, events_ttl, now);
@@ -335,6 +338,7 @@ impl Data {
             frozen: Timestamped::default(),
             timer_jobs: TimerJobs::default(),
             date_last_pinned: None,
+            gate,
         }
     }