+++ backend/canisters/user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Added `is_diamond_member` to `c2c_lookup_user`
+
 ### Removed
 
 - Removed temporary code needed for release ([#3375](https://github.com/open-ic/open-chat/pull/3375))